[UPDATE] merge eventRolesHandler with eventRoles
